Flatten class hierarchy for IR JSON-path nodes
Abstract classes `IrAccessor` and `IrMethod` are removed.
