--- core/trino-main/src/main/java/io/trino/json/ir/IrAccessor.java
@@ -1,45 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.trino.json.ir;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.trino.spi.type.Type;
-
-import java.util.Optional;
-
-import static java.util.Objects.requireNonNull;
-
-public abstract class IrAccessor
-        extends IrPathNode
-{
-    protected final IrPathNode base;
-
-    IrAccessor(IrPathNode base, Optional<Type> type)
-    {
-        super(type);
-        this.base = requireNonNull(base, "accessor base is null");
-    }
-
-    @Override
-    protected <R, C> R accept(IrJsonPathVisitor<R, C> visitor, C context)
-    {
-        return visitor.visitIrAccessor(this, context);
-    }
-
-    @JsonProperty
-    public IrPathNode getBase()
-    {
-        return base;
-    }
-}
